mill: implement z3.prepareOffline to download both macos and linux libs (#527)
this will be called in the nix build to ensure that the cached
dependencies are sufficient to build both mac and linux packages.

z3 library is also copied into the local build folder, rather than
copying it somewhere in the .cache directory.

adds Platform for handling platform (arch + os)-related things.

NativeDepsModule handles all the setting environment and library path.