add docker-helper.sh to lift examples using pac-nix docker image (#514)
this introduces the src/test/make/docker-helper.sh script which is
a (hopefully) developer-friendly method for lifting examples in a Docker
container provided by [pac-nix][].

to ensure reproducibility, this uses an x86_64-linux docker container
with cross-compiler provided by nix. for more details, see #288 which
the
docker-helper.sh is stolen from.

the usage is comprehensively documented in [readme].
the general flow is:
1. activate the docker-helper.sh environment: `eval $(docker-helper.sh
env)`.
   (you will need to re-run this if you close your terminal)
2. pull the docker image: `docker-helper.sh pull`
3. start the docker container: `docker-helper.sh start`
4. (optional) clean the directory you want to lift:
`make SUBDIRS=extraspec_incorrect/malloc_memcpy_strlen_memset_free
clean`
5. compile and lift the directory:
   `make SUBDIRS=extraspec_incorrect/malloc_memcpy_strlen_memset_free`
6. commit the results (eventually, we want to move the binaries into
an out-of-repo location. but until then, committing is probably fine.)

[pac-nix]: https://github.com/katrinafyi/pac-nix/
[readme]:
https://github.com/UQ-PAC/BASIL/blob/docker-helper/src/test/readme.md

this PR does not make any changes to the existing examples. it only
serves to make the Docker container available and accessible for people
who want to use it as a uniform lifting environment.

I do think that all examples should eventually be re-lifted using a
consistent method, but this will come later.