more deterministic ordering for IL serialise and sortProceduresRPO (#520)
in sortProceduresRPO, we use a TreeSet with a custom ordering to
enforce a deterministic ordering of walking when multiple options are
available. we also reason that the `seen` set is increasing
and so the `wl` is always decreasing. therefore, we can use `--=` and
avoid re-creating it every iteration. hopefully this is right.

in the IL's attribute printing, we sort all of the maps and sets before
printing. to do so, we introduce ProductOrdered which makes it easy
to define lexicographic ordering on a case class.

i also remove a problematic `Ordered[SpecGlobalOrAccess]` instance
which only sorted by address.

with this change, the attributes of the .il files are now consistent
between gtsrelf and oldrelf.

however, the block numbering and procedure ordering still changes
between old relf and gtsrelf