ci: sequence compile task before all other tasks
this lets other tasks use the cached output of the compile task,
so they don't need to re-build the project. previously, there was
no guaranteed sequencing so all the tasks would race to build and upload
the cache artifact.

this also lets us have one single point for producing the cache, so we
can make sure it contains all the test classes and semantic db data.