BREAKING! new uncurried form for `proc` DSL function
the dsl is rather troublesome because it wants to let the user specify
any subset of certain optional fields, while also allowing the blocks
to be given as a single iterable argument or as a variadic argument
list. previously, this was done with lots of overloads, but now we need
to add a new parameter `returnBlockLabel` and this is running up against
the limit of acceptable overload permutations. so, we need an approach
which lets one combine named parameters with defaults and variadic/list
arguments.

uncurried means it has two argument lists, and it needs to be used
like `proc(...)(...)`. this is the only way we can combine optional
named arguments with a varargs argument list.

for uses of `proc` which use the varargs form and supplied in/out
params, these have been changed to use the new uncurried form.

for certain uses of `proc`, we maintain some overloads to forward from
the older forms to the new form. namely, we have overloads for:
- the most simple label + block varargs form, and
- label + in/out params + block iterable form.

the ToScala generator has also been updated to emit the new forms where
necessary.