Loop invariant generation (#481)
Implements a loop invariant generation template. Abstract domains that
are able to encode their results as predicates (so boogie assertions)
can be used to get loop invariants by looking at the value of an
analysis result at loop headers. Adding a new analysis to the list is
simple (see `analysis/InvariantGenerator.scala`). The current way I've
set this up assumes the analysis is run using the intraproc worklist
solver, but it may be a good idea at some point to extend it to be
interprocedural in whatever way works.

```scala
  def addInvariantsToProc(procedure: Procedure) = {
    reversePostOrder(procedure)

    val intervals = DoubleIntervalDomain(Some(procedure))
    ForwardLoopInvariantGenerator(intervals, NarrowingWorklistSolver(intervals)).genAndAddInvariants(procedure)

    val gammas = MayGammaDomain(
      LatticeMap.TopMap(procedure.formalInParam.unsorted.map(v => (v, LatticeSet.FiniteSet(Set(v)))).toMap)
    )
    ForwardLoopInvariantGenerator(gammas, worklistSolver(gammas)).genAndAddInvariants(procedure)

    // Add more domains here
  }
```

A test case for generating constant loop bounds needed correct widening
and narrowing to give an optimal invariant, so the worklist solver was
extended to support this. I am unsure if I did this in the best way so
I'd appreciate it if someone could give their thoughts.