DSL inference of return block when omitted (Breaking change!) (#541)
this PR implements a `returnBlockLabel` parameter for the DSL's `proc` method and implements return block inference when this parameter is unspecified. if the parameter is specified (both Some and None), then inference is not done. this should make it appropriate for both ToScala/parsing which need to match the original IR identically, and for writing IR by hand where it's useful to infer return blocks if possible.

the dsl is rather troublesome because it wants to let the user specify
any subset of certain optional fields, while also allowing the blocks
to be given as a single iterable argument or as a variadic argument
list. previously, this was done with lots of overloads, but now we need
to add a new parameter `returnBlockLabel` and this is running up against
the limit of acceptable overload permutations. so, we need an approach
which lets one combine named parameters with defaults and variadic/list
arguments.

uncurried means it has two argument lists, and it needs to be used
like `proc(...)(...)`. this is the only way we can combine optional
named arguments with a varargs argument list.

this is a **breaking** change by changing the signature for the `proc`
method. you should keep this in mind when working with the DSL. the
ToScala generator has also been updated to emit the new forms where necessary.

for uses of `proc` which use the varargs form and supplied in/out
params, these have been changed to use the new uncurried form.

for certain uses of `proc`, we maintain some overloads to forward
the older forms to the new form. namely, we have overloads for:
- the most simple label + block varargs form, and
- label + in/out params + block iterable form.

going forwards, the uncurried form above will be the one to use for maximum flexibility. given its flexibility, it should work in most cases and i would hesitate to add additional overloads (one cases where it might be troublesome is the While-DSL which wants to pass a single `List[EventuallyBlock]` argument).