im crashing out...............
i want to be able to pass a value of ValueLattice which defines only a
subset of the operations, to ensure that only width-independent
functions are called on the provided `top` value.

F-bound polymorphism makes it hideously hard to do this because any
provided value must conform to `T <: ValueLattice[T]`, effectively
forcing it to be a subtype of T (which is TNum here).

if we were to make this restricted ValueLattice a subtype of TNum, we
also have big problems. first, since TNum already defines all the
methods, we would need to override them with implementations that throw.
secondly, this overriding needs to happen in TNum and would need to be
repeated if we wanted to do the same for other instances of ValueLattice
in the future.

this is deeply unsatisfying. honestly, idk what to do. i think that the
"good" solution would be to make it all type-classes. then, we would be
able to split up the "static" ops (width independent) with the
non-static ops that need to be called on existing values.

it would be very good to avoid the dual-purpose of ValueLattice as both
a "dictionary of functions" and as a "lattice value". this is the root
of problems here. when we pass around an abitrary (non-sound) value to
access the directory of functions, we want to guarantee that the caller
does not access the lattice value methods, because it would be unsound
to do so.