Retain order in nextBlocks and prevBlocks (#572)
loop headers for irreducible loops can be arbitrarily-chosen, and the
order of nextBlocks/prevBlocks affects this choice.

these sets are constructed from an order-retaining LinkedHashSet.
however, they need to be immutable so they were previously converted to
a HashSet which discards order. i've changed it to ListSet for now which
is fine for iteration but slow for membership checks. looking at the
current basil, i can't find any membership operations on these sets so
it's probably fine.

in future, we could roll our own immutable.LinkedHashSet if the need
arises.

i also added some code the the IrreducibleLoops test which lets you
observe the different return values from the loop finder. if the
assertion is replaced with set equals on the headers, then it will start
failing (some of the time).

after this change, this command goes through:
```bash
./mill run -i ../basilbench/bzip2/bzip2.gts --lifter --pc assert --simplify-tv-verify --main-procedure-name BZ2_decompress --trim-early --procedure-call-depth 1 --simplify-tv-dryrun -v
```

a different approach that doesn't quote work:
[0001-canonicalising-loop-representations.-HOWEVER.patch](https://github.com/user-attachments/files/22375999/0001-canonicalising-loop-representations.-HOWEVER.patch)