the thing i want to do is to "hoist" re-entry edges into the
outermost loop where they remain valid. that is, they should be
hoisted to the outermost loop containing "to" where "from" is not
internal to the loop.

i think to do this we need some notion of "self nodes", the nodes which
appear in /only/ a particular cycle and no other subcycles. is a "self
node" simply defined by iloop_header??????

much to think about