Replace irreducible loop analysis and transform (#578)
Replaces every line of IrreducibleLoops.scala. This reimplements the
algorithm in a way much closer to the paper. This new code does much
less explicit bookkeeping work. It replaces some old code that went
through the dfsp_pos with new code that does some kind of
topologically-ordered dynamic programming. In doing so, it is my hope
that it is more obviously correct.

Also:
- fixes certain bugs in the transform relating to topological order
- removes state machine
- adds mutability in data structures to make it closer to the paper
- fixes analysis and transform for nested irreducible loops

Closes https://github.com/UQ-PAC/BASIL/issues/410